Renamed to createComponent
