+++ plugins/synth-react/public/backstage.yaml
@@ -1,0 +1,5 @@
+<>: $B.<>
+toJS: $B.toJS
+log: $B.log
+Component: $B.Component
+resolve: $B.resolve
